minds API consolidation (4/n): remove superseded old app.py routes
Remove the old color / associate / disassociate / provider-toggle / recovery
(host-health, restart-system-interface, restart-host) / minds-lifecycle
(running, stop-hosts, stop-state-container) / backup-status routes + handlers +
now-dead helpers (app.py -919 lines); repoint the recovery page + e2e script to
v1; reconcile desktop/sharing response shapes with the frontend. Recovery
worker/health logic now lives only in workspace_recovery.py.

minds: 1532 passed / 6 skipped; latchkey 380; ratchets 56. WIP: removing the
last dead sharing/dismiss routes next.
