+++ apps/minds/changelog/mngr-more-minds-api.md
@@ -9,3 +9,5 @@ Continues the minds workspace-API work and merges the latest `main` into the bra
 - Adds the versioned recovery routes: `GET /api/v1/workspaces/<id>/health` returns the recovery diagnostics (probe list + dispatch tier), and `POST /api/v1/workspaces/<id>/restart` (`scope: services|host`, optional `host_already_stopped` for the host scope) dispatches a restart and returns an operation handle (`{operation_id, kind: "restart"}`) followed via the `/api/v1/workspaces/operations/<id>` (+`/logs`) resource, just like create and destroy. Restart status and a status-log stream are tracked in a new in-memory workspace-operation registry. The host-health probe and restart worker were extracted into a new `workspace_recovery` module so the API surface no longer depends on the desktop app module.
 
 - Moves the browser create flow onto the same `POST /api/v1/workspaces` endpoint agents use, so the two create paths can no longer drift. The create form now submits via a JS `fetch` (instead of a native HTML POST to `/create`); on success it navigates to `/creating/<operation_id>`, and validation errors are rendered inline from a structured JSON `400` (`{error, field}`, placed next to the offending input). The `/api/v1/workspaces` route gained the full create field surface the form needs: automatic `mind-N` naming when no name is given, an optional `color`, imbue_cloud template-version resolution, and the no-account imbue_cloud sign-up backstop (a `400` carrying `{redirect_url}`). The `/creating/<id>` page now polls the v1 operations resource (`GET /api/v1/workspaces/operations/<id>` + `/logs`) for status and live logs. The parallel create routes -- the `POST /create` form handler and the `/api/create-agent/<id>/status` + `/api/create-agent/<id>/logs` routes -- are removed; the `GET /create` form page and the `/creating/<id>` progress page remain.
+
+- Final cleanup pass: removes the now-superseded legacy desktop-client JSON routes from `app.py` (workspace color, account associate/disassociate, provider toggle, the recovery host-health/restart-system-interface/restart-host trio, the `/api/minds/running|stop-hosts|stop-state-container` quit-flow endpoints, and the batch `/api/backup-status`) along with their handlers and now-dead helpers; only the `/agents/<id>/recovery` page route remains (its JS now drives the versioned `GET /api/v1/workspaces/<id>/health` and `POST /api/v1/workspaces/<id>/restart` routes). Repoints the recovery page's restart dispatch to send a `{scope, host_already_stopped}` JSON body, and the `launch_to_msg_e2e` harness's create poll to `GET /api/v1/workspaces/operations/<id>`.
