+++ apps/minds/imbue/minds/test_ratchets.py
@@ -360,7 +360,7 @@ def test_prevent_inline_functions() -> None:
     # The added inline function is the ``record_loss`` helper nested in the
     # ported Sentry HTTP transport's ``_send_request`` (it closes over the
     # envelope being sent).
-    rc.check_inline_functions(_DIR, snapshot(10))
+    rc.check_inline_functions(_DIR, snapshot(9))
 
 
 def test_prevent_underscore_imports() -> None:
